requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0
opencv-python==4.8.1.78
easyocr==1.7.1
paddlepaddle==3.0.0
paddleocr==2.9.1
pdf2image==1.16.3
Pillow==10.2.0
python-dotenv==1.0.0
numpy<2.0.0